requirements.txt
Flask
Flask-JWT-Extended
Flask-Mail
Flask-Bcrypt
Flask-SQLAlchemy
itsdangerous
python-dotenv
